// super-admin/interface/type.tsx
export interface Sidebarmenu {
    title:string;
    path:string;
    icon: string;
}

export interface MerchantData {
  id: string;
  package: string;
  account: string;
  amount: string;
  customer: string;
  method:string;
  status: string;
  created: string;
}
export interface datas {
  id:string;
  title:string;
  desc:string;
}